requirements.txt
praw==7.8.1
python-dotenv==1.0.0
requests==2.31.0
groq==0.18.0
pytest==8.0.0
fastapi==0.109.2
sqlalchemy==2.0.27
httpx==0.27.0
supabase==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2